Applicazioni Distribuite e Mobili/Laboratorio/Django/djangotemplate/requirements.txt
django
python-dotenv
Pillow
djangorestframework
django-cors-headers